// src/app/new-document/models/question-type.interface.ts
export enum QuestionType {
    SINGLE_TEXTBOX = 'SINGLE_TEXTBOX',
    COMMENT_BOX = 'COMMENT_BOX',
    MULTIPLE_TEXTBOXES = 'MULTIPLE_TEXTBOXES',
    MULTIPLE_CHOICE = 'MULTIPLE_CHOICE',
    CHECKBOXES = 'CHECKBOXES',
    RANKING = 'RANKING',
    SLIDER = 'SLIDER',
    DATE_TIME = 'DATE_TIME',
    FILE_UPLOAD = 'FILE_UPLOAD',
    DROPDOWN = 'DROPDOWN',
    STAR_RATING = 'STAR_RATING'
}